To correct example code.
The comments in the example code stated the index.html fallback should kick when not SSR and when Production (exepct for dev).

The old example kicked in when not SSR or when process.env.PROD. This results on when dev mode in f.e. PWA mode in an error ("Uncaught non-precached-url: createHandlerBoundToURL('index.html') was called, but that URL is not precached. Please pass in a URL that is precached instead.) which prevents f.e. push event handlers to load correctly, making it hard to implement push notifications.